script_generator/requirements.txt
groq
python-dotenv
psycopg2-binary
sentence-transformers
numpy
pgvector